Direct connections now use RSA host key from power V3 file
When invoking SSH for a direct connection box now grabs the RSA host
key from the Box struct and writes it to `~/.ssh/box_known_hosts`
before invoking SSH with `-o UserKnownHostsFile=~/.ssh/box_known_hosts`.

There is potential for a race condition on that file if the user tries
to run two instances of box simultaneously. This is unlikely, but if
it becomes an issue we can investigate file locking for that file.

The `-o UserKnownHostsFile=...` option is supported by the oldest
SSH client version I can find in the office (OpenSSH_6.2p2) so I
suspect it should be fine everywhere.