Split `ls` output
The current output of `box ls` is a consistent list of space-separated
records, one per line. Except the first line. Assuming it was added
for human benefit, I suggest writing it to stderr - not sure if I'm
alone here, but I basically never run `box ls` except as part of a
shell pipeline.